Move radio functions to dedicated c and h files.
