+++ rev_b/firmware/LoFence/main.h
@@ -10,16 +10,8 @@
 
 #define DEBUG 1
 //#define FACTORY_RESET 1
-#define INTERVAL_SECONDS 60 * 72
-#define MEASURE_MS 15000
-
-#include "secret.h"
-// secret.h contains the required keys for OTAA
-// Copy from TTN Device Console.
-// Example:
-//   char *devEui = "FFFFFFFFFFFFFFFF"; // If set to all FF's here, we will use the Hardware EUI of the RN2483 module.
-//   char *appEui = "0000000000000000"; // The Things Stack uses all 00's by default.
-//   char *appKey = "";                 // Fill in your 32 character AppKey from the console here.
+#define INTERVAL_SECONDS 60 * 10
+#define MEASURE_MS 4000
 
 void adc_init();
