+++ rev_b/firmware/LoFence/radio.h
@@ -1,0 +1,17 @@
+/*
+ * radio.h
+ *
+ * Created: 2021/11/29 16:36:25
+ *  Author: jpmeijers
+ */ 
+
+
+#ifndef RADIO_H_
+#define RADIO_H_
+
+
+void transmit(uint16_t volt_bat, uint16_t volt_fence_plus, uint16_t volt_fence_minus);
+
+
+
+#endif /* RADIO_H_ */
